requirements.txt
# Core dependencies
numpy==1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning
torch>=2.0.0

# Hyperparameter Optimization
optuna>=3.3.0

# Configuration
pyyaml>=6.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.65.0

# Logging
colorlog>=6.7.0
